LFS-22 / LFS-115: Cleanup the serverside codebase Abstract NCIT specific methods independent of file type into an AbstractNCITParser class.
